app/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.7.0
Deprecated==1.2.13
Django==4.0.6
django-redis==5.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
packaging==21.3
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyparsing==3.0.9
pytz==2022.2.1
redis==4.3.4
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
wrapt==1.14.1
